requirements.txt
Flask==1.1.2  
Flask-Cors==3.0.10 
Flask-SQLAlchemy==2.5.1  
gunicorn==20.1.0 
python-dotenv==0.15.0 
requests==2.25.1
psycopg2-binary==2.8.6 